MET-5995 Moved part of the classes related with running flink job on a cluster by rest API to separate module.
